Remove erroneous generic cycle
$value is null if $values is defined, so this cycle could only
ever generate a ruleset with an empty value and hence invalid CSS.

The far better solution for generic classes without a value as part
of the selector is to add a key/value pair to the values map that
has '' (empty string) as key.

See commit 18ee2ecc27b48dfc33d502bb0de9031487cdbfbf for detais.